+++ src/main/java/org/prgrms/springbootbasic/entity/FixedAmountVoucher.java
@@ -6,20 +6,20 @@ public class FixedAmountVoucher implements Voucher {
 
     public static final String AMOUNT_MIN_RANGE_EXP_MSG = "amount는 0보다 작거나 같을 수 없습니다.";
     public static final String AMOUNT_MAX_RANGE_EXP_MSG = "amount는 100000보다 작아야 합니다.";
-    public static final long MIN_RANGE = 0L;
-    public static final long MAX_RANGE = 100000L;
+    public static final int MIN_RANGE = 0;
+    public static final int MAX_RANGE = 100000;
 
     private final UUID voucherId;
-    private final long amount;
+    private final int amount;
 
-    public FixedAmountVoucher(UUID voucherId, long amount) {
+    public FixedAmountVoucher(UUID voucherId, int amount) {
         validateAmountRange(amount);
 
         this.voucherId = voucherId;
         this.amount = amount;
     }
 
-    private void validateAmountRange(long amount) {
+    private void validateAmountRange(int amount) {
         if (amount <= MIN_RANGE) {
             throw new IllegalArgumentException(AMOUNT_MIN_RANGE_EXP_MSG);
         }
@@ -33,7 +33,7 @@ public class FixedAmountVoucher implements Voucher {
         return voucherId;
     }
 
-    public long getAmount() {
+    public int getAmount() {
         return amount;
     }
 }
